fix memory clobbering in synaptics code
I found there are errors in synaptics.c source file. One is overwriting memory
(other variables) and second is assign of synaptics internal numbers to
state->buttons (struct Gpm_Event).

This patch solves it. Array boundaries must be set to 8, because of for
cycle in tp_process_repeating_actions(). And assign of last_4_way_buttons
cannot be done to state->buttons. I have commented it out (and it seems to
work), but may be this is not the correct solution.

http://bugs.gentoo.org/36790

Signed-off-by: Mike Frysinger <vapier@gentoo.org>