libgpm: do not log debug messages
This change puts src/lib/report-lib.c more in line
with src/report.c by suppressing debug messages by
default (although here, the only way to change that
default is to change the code and recompile :).

This is particularly important to users of logcheck,
which for some reason considers any syslog message
containing the text "debug" to represent a
misconfigured system and a security risk.