For absolute positioned devices on /dev/input/eventX
Only tested with the Raspberry Pi "Official Touchscreen"(ft5406)
but may work with other touch screens and devices that report position
information in absolute coordiantes.

The options reverseX, reverseY, swapXY can be used to deal with screen
rotation. The fake3buttons uses the count of touchpoints to fake the
behavior of the 3 button mouse, with limited success.